Add Dagger 2 dependency; remove ActionManager
