+++ src/main/java/com/matzua/dagger/component/AppComponent.java
@@ -1,0 +1,9 @@
+package com.matzua.dagger.component;
+
+import com.matzua.App;
+import dagger.Component;
+
+@Component
+public interface AppComponent {
+    App getApp();
+}
